pyproject.toml
[tool.poetry]
name = "test-app-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
django = "^5.0.3"
google-cloud-storage = "^2.15.0"
google-api-python-client = "^2.121.0"
crcmod = "^1.7"
tqdm = "^4.66.2"
django-storages = {extras = ["google"], version = "^1.14.2"}
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
google-cloud-secret-manager = "^2.18.3"
google-auth = "^2.28.2"
ipykernel = "^6.29.3"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
